HW3-StreamApi fixes
Fixes according to comments
